requirements.txt
astroid==1.4.8
autopep8==1.2.4
beautifulsoup4==4.5.1
boto==2.45.0
braintree==3.33.0
conda==4.2.12
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.4
django-extensions==1.7.5
django-filter==1.0.1
django-modelcluster==2.0
django-storages==1.5.2
django-taggit==0.18.3
django-toolbelt==0.0.1
django-treebeard==4.1.0
django-webpack-loader==0.4.1
djangorestframework==3.5.3
gunicorn==19.6.0
html5lib==0.999999999
isort==4.2.5
lazy-object-proxy==1.2.2
Markdown==2.6.7
mccabe==0.5.3
Naked==0.1.31
pep8==1.7.0
Pillow==3.4.2
psycopg2==2.6.2
pycosat==0.6.1
pycrypto==2.6.1
pylint==1.6.4
pytz==2016.10
PyYAML==3.12
requests==2.11.1
ruamel-yaml===-VERSION
shellescape==3.4.1
six==1.10.0
static3==0.7.0
stripe==1.46.0
Unidecode==0.4.19
wagtail==1.7
webencodings==0.5
Willow==0.4
wrapt==1.10.8
